requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
pandas
python-dotenv
openai
tiktoken
